requirements.txt
flask
flask-cors
PyMuPDF
Pillow
requests
gunicorn
python-dotenv
google-genai
